Split the model store into multiple files
The source code for the model store is getting bigger, so splitting it
into multiple source files under a directory helps keeping it easier
to read.
