+++ ramalama/model_store/constants.py
@@ -1,0 +1,3 @@
+DIRECTORY_NAME_BLOBS = "blobs"
+DIRECTORY_NAME_REFS = "refs"
+DIRECTORY_NAME_SNAPSHOTS = "snapshots"
